move PyNumberSlots into protocol/number.rs
