+++ vm/src/types/slot.rs
@@ -8,8 +8,8 @@ use crate::{
     function::{Either, FromArgs, FuncArgs, OptionalArg, PyComparisonValue, PySetterValue},
     identifier,
     protocol::{
-        PyBuffer, PyIterReturn, PyMapping, PyMappingMethods, PyNumber, PyNumberBinaryFunc,
-        PyNumberBinaryOp, PyNumberMethods, PyNumberUnaryFunc, PySequence, PySequenceMethods,
+        PyBuffer, PyIterReturn, PyMapping, PyMappingMethods, PyNumber, PyNumberMethods,
+        PyNumberSlots, PySequence, PySequenceMethods,
     },
     vm::Context,
     AsObject, Py, PyObject, PyObjectRef, PyPayload, PyRef, PyResult, VirtualMachine,
@@ -106,120 +106,6 @@ impl PyTypeSlots {
 impl std::fmt::Debug for PyTypeSlots {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_str("PyTypeSlots")
-    }
-}
-
-#[derive(Default)]
-pub struct PyNumberSlots {
-    pub add: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub subtract: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub multiply: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub remainder: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub divmod: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub power: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub negative: AtomicCell<Option<PyNumberUnaryFunc>>,
-    pub positive: AtomicCell<Option<PyNumberUnaryFunc>>,
-    pub absolute: AtomicCell<Option<PyNumberUnaryFunc>>,
-    pub boolean: AtomicCell<Option<PyNumberUnaryFunc<bool>>>,
-    pub invert: AtomicCell<Option<PyNumberUnaryFunc>>,
-    pub lshift: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub rshift: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub and: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub xor: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub or: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub int: AtomicCell<Option<PyNumberUnaryFunc<PyRef<PyInt>>>>,
-    pub float: AtomicCell<Option<PyNumberUnaryFunc<PyRef<PyFloat>>>>,
-
-    pub right_add: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_subtract: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_multiply: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_remainder: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_divmod: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_power: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_lshift: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_rshift: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_and: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_xor: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_or: AtomicCell<Option<PyNumberBinaryFunc>>,
-
-    pub inplace_add: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_subtract: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_multiply: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_remainder: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_power: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_lshift: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_rshift: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_and: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_xor: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_or: AtomicCell<Option<PyNumberBinaryFunc>>,
-
-    pub floor_divide: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub true_divide: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_floor_divide: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_true_divide: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_floor_divide: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_true_divide: AtomicCell<Option<PyNumberBinaryFunc>>,
-
-    pub index: AtomicCell<Option<PyNumberUnaryFunc<PyRef<PyInt>>>>,
-
-    pub matrix_multiply: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub right_matrix_multiply: AtomicCell<Option<PyNumberBinaryFunc>>,
-    pub inplace_matrix_multiply: AtomicCell<Option<PyNumberBinaryFunc>>,
-}
-
-impl PyNumberSlots {
-    pub fn left_binary_op(&self, op_slot: PyNumberBinaryOp) -> Option<PyNumberBinaryFunc> {
-        use PyNumberBinaryOp::*;
-        match op_slot {
-            Add => self.add.load(),
-            Subtract => self.subtract.load(),
-            Multiply => self.multiply.load(),
-            Remainder => self.remainder.load(),
-            Divmod => self.divmod.load(),
-            Power => self.power.load(),
-            Lshift => self.lshift.load(),
-            Rshift => self.rshift.load(),
-            And => self.and.load(),
-            Xor => self.xor.load(),
-            Or => self.or.load(),
-            InplaceAdd => self.inplace_add.load(),
-            InplaceSubtract => self.inplace_subtract.load(),
-            InplaceMultiply => self.inplace_multiply.load(),
-            InplaceRemainder => self.inplace_remainder.load(),
-            InplacePower => self.inplace_power.load(),
-            InplaceLshift => self.inplace_lshift.load(),
-            InplaceRshift => self.inplace_rshift.load(),
-            InplaceAnd => self.inplace_and.load(),
-            InplaceXor => self.inplace_xor.load(),
-            InplaceOr => self.inplace_or.load(),
-            FloorDivide => self.floor_divide.load(),
-            TrueDivide => self.true_divide.load(),
-            InplaceFloorDivide => self.inplace_floor_divide.load(),
-            InplaceTrueDivide => self.inplace_true_divide.load(),
-            MatrixMultiply => self.matrix_multiply.load(),
-            InplaceMatrixMultiply => self.inplace_matrix_multiply.load(),
-        }
-    }
-
-    pub fn right_binary_op(&self, op_slot: PyNumberBinaryOp) -> Option<PyNumberBinaryFunc> {
-        use PyNumberBinaryOp::*;
-        match op_slot {
-            Add => self.right_add.load(),
-            Subtract => self.right_subtract.load(),
-            Multiply => self.right_multiply.load(),
-            Remainder => self.right_remainder.load(),
-            Divmod => self.right_divmod.load(),
-            Power => self.right_power.load(),
-            Lshift => self.right_lshift.load(),
-            Rshift => self.right_rshift.load(),
-            And => self.right_and.load(),
-            Xor => self.right_xor.load(),
-            Or => self.right_or.load(),
-            FloorDivide => self.right_floor_divide.load(),
-            TrueDivide => self.right_true_divide.load(),
-            MatrixMultiply => self.right_matrix_multiply.load(),
-            _ => None,
-        }
     }
 }
